Renamed buttons.js to buttonAlerts.js
- buttons.js felt too generic and might be confusing later
